pyproject.toml
[build-system]
requires = ["hatchling>=1.26.0"]
build-backend = "hatchling.build"

[project]
name = "difftriage"
version = "0.1.0"
description = "PR risk gate for AI-generated code: diff scoring, test impact checks, and merge policy enforcement."
readme = "README.md"
requires-python = ">=3.10"
license = { text = "MIT" }
authors = [{ name = "DiffTriage Contributors" }]
dependencies = [
  "pyyaml>=6.0",
  "typer>=0.12.0",
]

[project.scripts]
difftriage = "difftriage.cli:app"

[tool.pytest.ini_options]
testpaths = ["tests"]